fix remarks

+++ hw03-generics/src/main/java/ru/otus/generics/model/Apple.java
@@ -1,7 +1,7 @@
 package ru.otus.generics.model;
 
 public class Apple extends Fruit {
-    public Apple(int weight) {
+    public Apple(float weight) {
         super(weight);
     }
 }
