+++ hw03-generics/src/main/java/ru/otus/generics/model/Fruit.java
@@ -1,17 +1,17 @@
 package ru.otus.generics.model;
 
 public abstract class Fruit {
-    private int weight;
+    private float weight;
 
-    public int getWeight() {
+    public float getWeight() {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(float weight) {
         this.weight = weight;
     }
 
-    public Fruit(int weight) {
+    public Fruit(float weight) {
         this.weight = weight;
     }
 }
